+++ src/fsd_entities/auth/model/hooks/index.ts
@@ -5,3 +5,6 @@ export * from './useFindEmailForm';
 export * from './useFindPasswordForm';
 export * from './useV2Form';
 export * from './useAcademicRecordForm';
+export * from './useSignUpForm';
+export * from './useAdmissionForm';
+export * from './useVerification';
