+++ src/fsd_entities/auth/model/hooks/useAdmissionForm.ts
@@ -1,15 +1,11 @@
-import { useRouter } from 'next/navigation';
-
-import { useMutation } from '@tanstack/react-query';
-import { set, useForm } from 'react-hook-form';
+import { useForm } from 'react-hook-form';
 import toast from 'react-hot-toast';
 
 import { useUserProfile } from '@/fsd_entities/user/model';
 
-import { submitAdmissionsApplication } from '../../api/post';
+import { usePostAdmission } from '../queries';
 
 export const useAdmissionForm = () => {
-  const router = useRouter();
   const { data: userInfo } = useUserProfile();
 
   const email = userInfo?.email || '';
@@ -20,25 +16,14 @@ export const useAdmissionForm = () => {
     formState: { errors },
   } = useForm<User.AdmissionCreateRequestDto>({ mode: 'onBlur' });
 
-  const mutation = useMutation({
-    mutationFn: submitAdmissionsApplication,
-    onSuccess: () => {
-      toast.success('가입 신청서 제출이 완료되었습니다!');
-      setTimeout(() => {
-        router.push('/auth/authorization');
-      }, 500);
-    },
-    onError: (error: any) => {
-      toast.error('가입 신청서 제출 실패: ' + (error || '오류가 발생했습니다.'));
-    },
-  });
+  const submitAdmission = usePostAdmission();
 
   const onSubmit = (data: User.AdmissionCreateRequestDto) => {
     if (!data.attachImage) {
       toast.error('이미지를 첨부해주세요.');
       return;
     }
-    mutation.mutate(data);
+    submitAdmission.mutate(data);
   };
 
   const onInvalid = () => {
