+++ src/fsd_entities/auth/model/queries/index.ts
@@ -1,5 +1,4 @@
 export * from './useFindPassword';
 export * from './useFindId';
-export * from './useVerification';
-export * from './useSignUpForm';
-export * from './useAdmissionForm';
+export * from './usePostAdmission';
+export * from './usePostSignUp';
